+++ feature/transactions/src/main/kotlin/nz/mikhailov/motiv/feature/transactions/ui/model/RewardUIO.kt
@@ -17,43 +17,43 @@ import nz.mikhailov.motiv.feature.transactions.ui.model.RewardUIO.Study
 import nz.mikhailov.motiv.feature.transactions.ui.model.RewardUIO.Unknown
 
 sealed interface RewardUIO {
-    val amount: Int
+    val amount: Double
     val icon: ImageVector
     val iconString: String
         get() = icon.name
     val actionDescription: String
 
     data class Exercise(
-        override val amount: Int,
+        override val amount: Double,
         override val icon: ImageVector = Icons.Filled.FitnessCenter,
         override val actionDescription: String = "Add exercise reward"): RewardUIO
 
     data class Study(
-        override val amount: Int,
+        override val amount: Double,
         override val icon: ImageVector = Icons.Filled.School,
         override val actionDescription: String = "Add study reward"): RewardUIO
 
     data class Code(
-        override val amount: Int,
+        override val amount: Double,
         override val icon: ImageVector = Icons.Filled.DeveloperMode,
         override val actionDescription: String = "Add coding reward"): RewardUIO
 
     data class NoJunkFood(
-        override val amount: Int,
+        override val amount: Double,
         override val icon: ImageVector = Icons.Filled.NoFood,
         override val actionDescription: String = "Add no junk food reward"): RewardUIO
 
     data class Unknown(
-        override val amount: Int,
+        override val amount: Double,
         override val icon: ImageVector = Icons.Filled.Help,
         override val actionDescription: String = ""): RewardUIO
 }
 
 fun Transaction.toRewardUIO() = when (activity) {
-    Exercise(0).iconString -> Exercise(amount)
-    Study(0).iconString -> Study(amount)
-    Code(0).iconString -> Code(amount)
-    NoJunkFood(0).iconString -> NoJunkFood(amount)
+    Exercise(0.0).iconString -> Exercise(amount)
+    Study(0.0).iconString -> Study(amount)
+    Code(0.0).iconString -> Code(amount)
+    NoJunkFood(0.0).iconString -> NoJunkFood(amount)
     else -> Unknown(amount)
 }
