+++ feature/transactions/src/main/kotlin/nz/mikhailov/motiv/feature/transactions/ui/model/TransactionsUIO.kt
@@ -3,11 +3,11 @@ package nz.mikhailov.motiv.feature.transactions.ui.model
 import nz.mikhailov.motiv.feature.transactions.business.model.Transaction
 
 data class TransactionsUIO(
-    val balance: Int,
+    val balance: Double,
     val transactions: List<TransactionUIO>,
 )
 
 fun List<Transaction>.toUIO() = TransactionsUIO(
-    balance = firstOrNull()?.balance ?: 0,
+    balance = firstOrNull()?.balance ?: 0.0,
     transactions = map(Transaction::toUIO),
 )
